pyproject.toml
[tool.black]
line-length = 88
skip-string-normalization = true

[tool.poetry]
name = "waldur-mastermind"
version = "0.0.0"
description = "Waldur MasterMind is a hybrid cloud orchestrator."
authors = ["OpenNode Team <info@opennodecloud.com>"]
license = "MIT"
homepage = "https://waldur.com"
documentation = "http://docs.waldur.com"
classifiers=[
  "Framework :: Django",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: MIT License",
]

[tool.poetry.dependencies]
python = "^3.8"
apache-libcloud = "2.2.1"
azure-mgmt-compute = "5.0.0"
azure-mgmt-network = "2.0.0"
azure-mgmt-nspkg = "3.0.2"
azure-mgmt-rdbms = "1.2.0"
azure-mgmt-resource = "2.0.0"
azure-mgmt-sql = "0.5.1"
azure-mgmt-storage = "2.0.0"
celery = "4.3.0"
croniter = "0.3.5"
defusedxml = "0.6.0"
django = "2.2.13"
django-admin-tools = "0.9.0"
django-auth-ldap = "2.0.0"
django-import-export = "2.3.0"
django-filter = "2.2"
django-fluent-dashboard = "1.0.1"
django-fsm = "2.6.1"
django-jsoneditor = "0.0.12"
django-model-utils = "3.2.0"
django-redis-cache = "2.0.0"
django-rest-swagger = "2.1.2"
django-reversion = "2.0.8"
django-taggit = "1.1.0"
djangorestframework = "3.10.2"
djangosaml2 = "1.0.5"
docker = "4.1.0"
hiredis = "0.2.0"
influxdb = "5.0.0"
iptools = "0.6.1"
jira = "2.0.0"
lxml = "4.2.1"
paypalrestsdk = "1.13.1"
pbr = "4.0.4"
pdfkit = "0.6.1"
Pillow = "7.2.0"
prettytable = "0.7.2"
psycopg2-binary = "2.8.4"
pycountry = "1.20"
PyJWT = "1.5.3"
pyOpenSSL = "18.0.0"
pysaml2 = "6.5.1"
python-cinderclient = "3.6.0"
python-dateutil = "2.7.3"
python-digitalocean = "1.16.0"
python-glanceclient = "2.11.1"
python-keystoneclient = "3.16.0"
python-ldap = "3.1.0"
python-neutronclient = "6.12.0"
python-novaclient = "10.3.0"
pyvat = "1.3.15"
pyvmomi = "6.7.1"
PyYAML = "5.1.2"
pyzabbix = "0.7.4"
redis = "3.2.1"
requests = "2.20.0"
sentry-sdk = "0.13.1"
sqlparse = "0.2.4"
python-freeipa = "0.2.5"
django-modeltranslation = "^0.15"
django-storages = "^1.9.1"
boto3 = "^1.14.18"
django-axes = "^5.4.3"
geopy = "^2.0.0"
google-api-python-client = "^1.12.5"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
django-health-check = "^3.16.2"
bleach = "^3.3.0"
django-dbtemplates = {git = "https://github.com/jazzband/django-dbtemplates.git"}
ansible-waldur-module = {git = "https://github.com/opennode/ansible-waldur-module.git", rev = "develop"}

[tool.poetry.dev-dependencies]
pytest = "3.6.2"
pytest-cov = "2.5.1"
pytest-django = "3.3.2"
responses = "0.9.0"
ddt = "1.0.3"
factory-boy = "2.4.1"
freezegun = "0.3.7"
mock-django = "0.6.9"
SQLAlchemy = "1.3.0"
django-cors-headers = "^3.2.1"
attrs = "18.1.0"
pre-commit = "^2.1.1"
pytest-split-tests = "^1.0.9"

[tool.poetry.plugins.console_scripts]
waldur = "waldur_core.server.manage:main"

[tool.poetry.plugins.waldur_extensions]
waldur_analytics = "waldur_mastermind.analytics.extension:AnalyticsExtension"
waldur_auth_bcc = "waldur_auth_bcc.extension:AuthBCCExtension"
waldur_auth_saml2 = "waldur_auth_saml2.extension:SAML2Extension"
waldur_auth_social = "waldur_auth_social.extension:AuthSocialExtension"
waldur_auth_valimo = "waldur_auth_valimo.extension:AuthValimoExtension"
waldur_aws = "waldur_aws.extension:AWSExtension"
waldur_azure = "waldur_azure.extension:AzureExtension"
waldur_billing = "waldur_mastermind.billing.extension:BillingExtension"
waldur_booking = "waldur_mastermind.booking.extension:BookingExtension"
waldur_digitalocean = "waldur_digitalocean.extension:DigitalOceanExtension"
waldur_freeipa = "waldur_freeipa.extension:FreeIPAExtension"
waldur_geo_ip = "waldur_geo_ip.extension:GeoIPExtension"
waldur_invoices = "waldur_mastermind.invoices.extension:InvoicesExtension"
waldur_jira = "waldur_jira.extension:JiraExtension"
waldur_marketplace = "waldur_mastermind.marketplace.extension:MarketplaceExtension"
waldur_marketplace_azure = "waldur_mastermind.marketplace_azure.extension:MarketplaceAzureExtension"
waldur_marketplace_checklist = "waldur_mastermind.marketplace_checklist.extension:MarketplaceChecklistExtension"
waldur_marketplace_openstack = "waldur_mastermind.marketplace_openstack.extension:MarketplaceOpenStackExtension"
waldur_marketplace_rancher = "waldur_mastermind.marketplace_rancher.extension:MarketplaceRancherExtension"
waldur_marketplace_script = "waldur_mastermind.marketplace_script.extension:MarketplaceScriptExtension"
waldur_marketplace_slurm = "waldur_mastermind.marketplace_slurm.extension:MarketplaceSlurmExtension"
waldur_marketplace_support = "waldur_mastermind.marketplace_support.extension:MarketplaceSupportExtension"
waldur_marketplace_vmware = "waldur_mastermind.marketplace_vmware.extension:MarketplaceVMwareExtension"
waldur_marketplace_remote = "waldur_mastermind.marketplace_remote.extension:MarketplaceRemoteExtension"
waldur_openstack = "waldur_openstack.openstack.extension:OpenStackExtension"
waldur_openstack_tenant = "waldur_openstack.openstack_tenant.extension:OpenStackTenantExtension"
waldur_notifications = "waldur_mastermind.notifications.extension:NotificationsExtension"
waldur_paypal = "waldur_paypal.extension:PayPalExtension"
waldur_rancher = "waldur_rancher.extension:RancherExtension"
waldur_slurm = "waldur_slurm.extension:SlurmExtension"
waldur_support = "waldur_mastermind.support.extension:SupportExtension"
waldur_vmware = "waldur_vmware.extension:VMwareExtension"
waldur_zabbix = "waldur_zabbix.extension:ZabbixExtension"
waldur_pid = "waldur_pid.extension:PIDExtension"
waldur_google = "waldur_mastermind.google.extension:GoogleExtension"

[build-system]
requires = ["poetry=1.0.9"]
build-backend = "poetry.masonry.api"
